uiux_bot/requirements.txt
python-telegram-bot==20.7
openai==1.6.1
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
pytz==2023.3
aiohttp==3.9.1
tenacity==8.2.3
pydantic==2.5.2
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
flake8==6.1.0
black==23.11.0 